backend/requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.2
cryptography==46.0.3
Deprecated==1.3.1
ecdsa==0.19.1
Faker==40.1.2
fastapi==0.128.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
passlib==1.7.4
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
python-jose==3.5.0
python-json-logger==4.0.0
redis==7.1.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wrapt==2.0.1
